// src/lib/api/generated/model/emailSendLog.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { EmailSendLogStatusEnum } from './emailSendLogStatusEnum';

export interface EmailSendLog {
  readonly id: string;
  readonly campaign_recipient: string;
  readonly tenant: string;
  readonly email_template: string;
  /** @nullable */
  readonly campaign_step: string | null;
  readonly email_to: string;
  readonly email_subject: string;
  readonly status: EmailSendLogStatusEnum;
  readonly aws_message_id: string;
  readonly error_message: string;
  readonly retry_count: number;
  /** @nullable */
  readonly sent_at: string | null;
  readonly created_at: string;
  readonly updated_at: string;
}