// src/lib/api/generated/model/emailSendLogStatusEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `PENDING` - Pending
* `SENT` - Sent
* `DELIVERED` - Delivered
* `FAILED` - Failed
* `BOUNCED` - Bounced
 */
export type EmailSendLogStatusEnum = typeof EmailSendLogStatusEnum[keyof typeof EmailSendLogStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EmailSendLogStatusEnum = {
  PENDING: 'PENDING',
  SENT: 'SENT',
  DELIVERED: 'DELIVERED',
  FAILED: 'FAILED',
  BOUNCED: 'BOUNCED',
} as const;